+++ Project/library/src/Core/Processer.cpp
@@ -2,7 +2,7 @@
 
 
 
-int Processer::truncate(int value)
+int Processer::truncate(double value)
 {
 	if (value < 0) return 0;
 	if (value > 255) return 255;
